Move cLogStats into ISLogStats.cpp and declare them as static members of cISLogger, not dynamically allocated
