+++ CMakeLists.txt
@@ -41,6 +41,7 @@ add_library(InertialSense
   src/ISFileManager.cpp
   src/ISLogFile.cpp
   src/ISLogger.cpp
+  src/ISLogStats.cpp
   src/ISMatrix.c
   src/ISPose.c
   src/ISSerialPort.cpp
